new update

+++ fast_food/database/migrations/2022_05_01_144033_create_don_hangs_table.php
@@ -16,7 +16,7 @@ class CreateDonHangsTable extends Migration
         Schema::create('don_hangs', function (Blueprint $table) {
             $table->id();
             $table->dateTime('ngay_lap_dh');
-            $table->double('tong_tien');
+            $table->decimal('tong_tien', 9, 3);
             $table->unsignedBigInteger('trang_thai_don_hang_id')->nullable(); //khoá ngoại
             $table->unsignedBigInteger('nguoi_giao_hang_id')->nullable(); //khoá ngoại
             $table->unsignedBigInteger('user_id')->nullable(); //khoá ngoại
